+++ ampltools/CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 0.4.7 - 2023-03-02
 - Add modules.activate and activate command.
 - Add modules.preload to load modules silently.
+- Add modules.unload to unload modules.
 - Add modules.run.
 
 ## 0.4.6 - 2023-01-24